feat: Implement course creation and management UI
This commit introduces a new user interface for course creation and management in the admin dashboard.

The existing course management functionality within the `AdminDashboard.tsx` component has been refactored into three new, more modular components:

- `CourseManagement.tsx`: A container component that manages the state and logic for course operations.
- `CourseForm.tsx`: A form component for creating and editing courses.
- `CourseList.tsx`: A component that displays the list of courses and provides actions for editing and deleting them.

This refactoring improves code organization, separates concerns, and makes the course management feature more maintainable. The `AdminDashboard.tsx` is now cleaner and delegates all course-related UI and logic to the new `CourseManagement` component.
