+++ src/pages/AdminDashboard.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-import { Users, BookOpen, UserCheck, GraduationCap, Plus, CreditCard as Edit, Trash, Eye } from 'lucide-react';
-import { useForm } from 'react-hook-form';
+import { Users, BookOpen, UserCheck, GraduationCap, Eye } from 'lucide-react';
+import CourseManagement from '../components/CourseManagement';
 
 const API_URL = import.meta.env.VITE_API_URL || 'http://localhost:5000/api';
 
@@ -48,10 +48,6 @@ const AdminDashboard: React.FC = () => {
   const [selectedCourseProgress, setSelectedCourseProgress] = useState<any>(null);
   const [statistics, setStatistics] = useState<any>(null);
   const [loading, setLoading] = useState(false);
-  const [showCourseForm, setShowCourseForm] = useState(false);
-  const [editingCourse, setEditingCourse] = useState<Course | null>(null);
-
-  const { register, handleSubmit, reset, setValue } = useForm();
 
   useEffect(() => {
     fetchData();
@@ -178,41 +174,6 @@ const AdminDashboard: React.FC = () => {
     }
   };
 
-  const onSubmitCourse = async (data: any) => {
-    try {
-      if (editingCourse) {
-        await axios.put(`${API_URL}/admin/courses/${editingCourse._id}`, data);
-      } else {
-        await axios.post(`${API_URL}/admin/courses`, data);
-      }
-      
-      reset();
-      setShowCourseForm(false);
-      setEditingCourse(null);
-      fetchCourses();
-    } catch (error) {
-      console.error('Error saving course:', error);
-    }
-  };
-
-  const editCourse = (course: Course) => {
-    setEditingCourse(course);
-    setValue('title', course.title);
-    setValue('description', course.description);
-    setValue('isPublished', course.isPublished);
-    setShowCourseForm(true);
-  };
-
-  const deleteCourse = async (courseId: string) => {
-    if (confirm('Are you sure you want to delete this course?')) {
-      try {
-        await axios.delete(`${API_URL}/admin/courses/${courseId}`);
-        fetchCourses();
-      } catch (error) {
-        console.error('Error deleting course:', error);
-      }
-    }
-  };
 
   const StatCard: React.FC<{ title: string; value: number; icon: React.ReactNode; color: string }> = ({ title, value, icon, color }) => (
     <div className="bg-white p-6 rounded-lg shadow-lg">
@@ -490,139 +451,11 @@ const AdminDashboard: React.FC = () => {
 
         {/* Courses Tab */}
         {activeTab === 'courses' && (
-          <div className="space-y-6">
-            {/* Course Form */}
-            {showCourseForm && (
-              <div className="bg-white rounded-lg shadow-lg p-6">
-                <h3 className="text-lg font-semibold text-gray-900 mb-4">
-                  {editingCourse ? 'Edit Course' : 'Create New Course'}
-                </h3>
-                <form onSubmit={handleSubmit(onSubmitCourse)} className="space-y-4">
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-1">Title</label>
-                    <input
-                      {...register('title', { required: true })}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
-                      placeholder="Course title"
-                    />
-                  </div>
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-1">Description</label>
-                    <textarea
-                      {...register('description', { required: true })}
-                      rows={3}
-                      className="w-full px-3 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-blue-500"
-                      placeholder="Course description"
-                    />
-                  </div>
-                  <div className="flex items-center">
-                    <input
-                      {...register('isPublished')}
-                      type="checkbox"
-                      className="h-4 w-4 text-blue-600 border-gray-300 rounded"
-                    />
-                    <label className="ml-2 block text-sm text-gray-700">Published</label>
-                  </div>
-                  <div className="flex space-x-4">
-                    <button
-                      type="submit"
-                      className="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700"
-                    >
-                      {editingCourse ? 'Update' : 'Create'} Course
-                    </button>
-                    <button
-                      type="button"
-                      onClick={() => {
-                        setShowCourseForm(false);
-                        setEditingCourse(null);
-                        reset();
-                      }}
-                      className="bg-gray-300 text-gray-700 px-4 py-2 rounded-lg hover:bg-gray-400"
-                    >
-                      Cancel
-                    </button>
-                  </div>
-                </form>
-              </div>
-            )}
-
-            {/* Courses List */}
-            <div className="bg-white rounded-lg shadow-lg">
-              <div className="p-6 border-b flex justify-between items-center">
-                <h3 className="text-lg font-semibold text-gray-900">Courses</h3>
-                <button
-                  onClick={() => setShowCourseForm(true)}
-                  className="bg-blue-600 text-white px-4 py-2 rounded-lg hover:bg-blue-700 flex items-center space-x-2"
-                >
-                  <Plus className="h-4 w-4" />
-                  <span>New Course</span>
-                </button>
-              </div>
-              <div className="overflow-x-auto">
-                <table className="min-w-full divide-y divide-gray-200">
-                  <thead className="bg-gray-50">
-                    <tr>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Course</th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Instructor</th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Status</th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Enrollments</th>
-                      <th className="px-6 py-3 text-left text-xs font-medium text-gray-500 uppercase tracking-wider">Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody className="bg-white divide-y divide-gray-200">
-                    {courses.map((course) => (
-                      <tr key={course._id}>
-                        <td className="px-6 py-4 whitespace-nowrap">
-                          <div>
-                            <div className="text-sm font-medium text-gray-900">{course.title}</div>
-                            <div className="text-sm text-gray-500 line-clamp-2">{course.description}</div>
-                          </div>
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                          {course.instructor.name}
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap">
-                          <span className={`inline-flex px-2 py-1 text-xs font-semibold rounded-full ${
-                            course.isPublished ? 'bg-green-100 text-green-800' : 'bg-gray-100 text-gray-800'
-                          }`}>
-                            {course.isPublished ? 'Published' : 'Draft'}
-                          </span>
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm text-gray-900">
-                          {course.enrollmentCount}
-                        </td>
-                        <td className="px-6 py-4 whitespace-nowrap text-sm font-medium">
-                          <div className="flex space-x-2">
-                            <button
-                              onClick={() => editCourse(course)}
-                              className="text-blue-600 hover:text-blue-900"
-                              title="Edit Course"
-                            >
-                              <Edit className="h-4 w-4" />
-                            </button>
-                            <button
-                              onClick={() => fetchCourseProgress(course._id)}
-                              className="text-green-600 hover:text-green-900"
-                              title="View Progress"
-                            >
-                              <Eye className="h-4 w-4" />
-                            </button>
-                            <button
-                              onClick={() => deleteCourse(course._id)}
-                              className="text-red-600 hover:text-red-900"
-                              title="Delete Course"
-                            >
-                              <Trash className="h-4 w-4" />
-                            </button>
-                          </div>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </table>
-              </div>
-            </div>
-          </div>
+          <CourseManagement
+            courses={courses}
+            fetchCourses={fetchCourses}
+            onViewCourseProgress={fetchCourseProgress}
+          />
         )}
 
         {/* Enrollments Tab */}
